+++ trixnity-messenger-compose-view/src/androidMain/kotlin/de/connect2x/trixnity/messenger/compose/view/previews/RoomListElementPreview.kt
@@ -21,6 +21,7 @@ private fun RoomListElementPreview() {
             roomListElementViewModel = roomElementViewModel,
             roomListViewModel = PreviewRoomListViewModel(),
             index = 0,
+            showRoomTime = true
         )
     }
 }