+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - SDK: TextFieldViewModel: allow for input validation
 - SDK: The floating 'create new chat or group' button is now part of the DI
 - UI: Info that pre-join history is unavailable. That is in the timeline for the joined user and for the inviter when adding a member
+- UI: Context menu for room list which currently includes the option to mark as unread
 
 ### Changed
 
@@ -28,6 +29,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - UI: Correctly format markdown based privacy and imprint
 - UI: Correctly render chat bubble edge for own messages
 - UI,Android: Keep Wizard fullscreen after keyboard popup
+- SDK: Mark rooms as read when they are marked as unread without having unread events
 
 ### Security